+++ src/MoonrakerSharpWebApi/MoonrakerClient.cs
@@ -98,71 +98,6 @@ namespace AndreasReitberger.API.Moonraker
 
         #endregion
 
-        #region Printer       
-
-        [ObservableProperty]
-        [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-        double liveVelocity = 0;
-
-        [ObservableProperty]
-        [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-        double liveExtruderVelocity = 0;
-
-        #endregion
-
-        #region RemotePrinters
-        /*
-        [ObservableProperty]
-        [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-        ObservableCollection<KlipperDatabaseRemotePrinter> printers = new();
-        partial void OnPrintersChanged(ObservableCollection<KlipperDatabaseRemotePrinter> value)
-        {
-            OnKlipperRemotePrinterChanged(new KlipperRemotePrintersChangedEventArgs()
-            {
-                NewPrinters = value,
-                SessonId = SessionId,
-                CallbackId = -1,
-                Token = !string.IsNullOrEmpty(UserToken) ? UserToken : ApiKey,
-            });
-        }
-        */
-        #endregion
-
-        #region Files
-        /*
-        [ObservableProperty]
-        [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-        ObservableCollection<KlipperFile> files = new();
-        partial void OnFilesChanged(ObservableCollection<KlipperFile> value)
-        {
-            OnKlipperFilesChanged(new KlipperFilesChangedEventArgs()
-            {
-                NewFiles = value,
-                SessonId = SessionId,
-                CallbackId = -1,
-                Token = !string.IsNullOrEmpty(UserToken) ? UserToken : ApiKey,
-            });
-        }
-        */
-
-        [ObservableProperty]
-        [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-        ObservableCollection<KlipperDirectory> availableDirectories = new();
-        partial void OnAvailableDirectoriesChanged(ObservableCollection<KlipperDirectory> value)
-        {
-            /*
-            OnKlipperFilesChanged(new KlipperFilesChangedEventArgs()
-            {
-                NewFiles = value,
-                SessonId = SessionId,
-                CallbackId = -1,
-                Token = !string.IsNullOrEmpty(UserToken) ? UserToken : API,
-            });
-            */
-        }
-
-        #endregion
-
         #region Jobs
         /*
         [ObservableProperty]
@@ -292,20 +227,6 @@ namespace AndreasReitberger.API.Moonraker
 
         [ObservableProperty]
         [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-        KlipperPrinterStateMessageResult printerInfo;
-        partial void OnPrinterInfoChanged(KlipperPrinterStateMessageResult value)
-        {
-            OnKlipperPrinterInfoChanged(new KlipperPrinterInfoChangedEventArgs()
-            {
-                NewPrinterInfo = value,
-                SessonId = SessionId,
-                CallbackId = -1,
-            });
-            UpdatePrinterInfo(value);
-        }
-
-        [ObservableProperty]
-        [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
         KlipperGcodeMetaResult gcodeMeta;
         partial void OnGcodeMetaChanged(KlipperGcodeMetaResult value)
         {
@@ -348,7 +269,6 @@ namespace AndreasReitberger.API.Moonraker
 
         [ObservableProperty]
         [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-#if ConcurrentDictionary
         ConcurrentDictionary<string, KlipperStatusTemperatureSensor> temperatureSensors = new();
         partial void OnTemperatureSensorsChanged(ConcurrentDictionary<string, KlipperStatusTemperatureSensor> value)
         {
@@ -377,40 +297,9 @@ namespace AndreasReitberger.API.Moonraker
                 });
             }
         }
-#else
-        Dictionary<string, KlipperStatusTemperatureSensor> temperatureSensors = new();
-        partial void OnTemperatureSensorsChanged(Dictionary<string, KlipperStatusTemperatureSensor> value)
-        {
-            if (_enableCooldown)
-            {
-                if (_cooldownTemperatureSensor > 0)
-                    _cooldownTemperatureSensor--;
-                else
-                {
-                    _cooldownTemperatureSensor = _cooldownFallback;
-                    OnKlipperTemperatureSensorStatesChanged(new KlipperTemperatureSensorStatesChangedEventArgs()
-                    {
-                        TemperatureStates = value,
-                        SessonId = SessionId,
-                        CallbackId = -1,
-                    });
-                }
-            }
-            else
-            {
-                OnKlipperTemperatureSensorStatesChanged(new KlipperTemperatureSensorStatesChangedEventArgs()
-                {
-                    TemperatureStates = value,
-                    SessonId = SessionId,
-                    CallbackId = -1,
-                });
-            }
-        }
-#endif
 
         [ObservableProperty]
         [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-#if ConcurrentDictionary
         ConcurrentDictionary<string, double?> cpuUsage = new();
         partial void OnCpuUsageChanged(ConcurrentDictionary<string, double?> value)
         {
@@ -440,41 +329,9 @@ namespace AndreasReitberger.API.Moonraker
                 });
             }
         }
-#else
-        Dictionary<string, double?> cpuUsage = new();
-        partial void OnCpuUsageChanged(Dictionary<string, double?> value)
-        {
-            // WebSocket is updating this property in a high frequency, so a cooldown can be enabled
-            if (_enableCooldown)
-            {
-                if (_cooldownCpuUsage > 0)
-                    _cooldownCpuUsage--;
-                else
-                {
-                    _cooldownCpuUsage = _cooldownFallback;
-                    OnKlipperServerCpuUsageChanged(new KlipperCpuUsageChangedEventArgs()
-                    {
-                        CpuUsage = value,
-                        SessonId = SessionId,
-                        CallbackId = -1,
-                    });
-                }
-            }
-            else
-            {
-                OnKlipperServerCpuUsageChanged(new KlipperCpuUsageChangedEventArgs()
-                {
-                    CpuUsage = value,
-                    SessonId = SessionId,
-                    CallbackId = -1,
-                });
-            }
-        }
-#endif
 
         [ObservableProperty]
         [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-#if ConcurrentDictionary
         ConcurrentDictionary<string, long?> systemMemory = new();
         partial void OnSystemMemoryChanged(ConcurrentDictionary<string, long?> value)
         {
@@ -504,37 +361,6 @@ namespace AndreasReitberger.API.Moonraker
                 });
             }
         }
-#else
-        Dictionary<string, long?> systemMemory = new();
-        partial void OnSystemMemoryChanged(Dictionary<string, long?> value)
-        {
-            // WebSocket is updating this property in a high frequency, so a cooldown can be enabled
-            if (_enableCooldown)
-            {
-                if (_cooldownSystemMemory > 0)
-                    _cooldownSystemMemory--;
-                else
-                {
-                    _cooldownSystemMemory = _cooldownFallback;
-                    OnKlipperServerSystemMemoryChanged(new KlipperSystemMemoryChangedEventArgs()
-                    {
-                        SystemMemory = value,
-                        SessonId = SessionId,
-                        CallbackId = -1,
-                    });
-                }
-            }
-            else
-            {
-                OnKlipperServerSystemMemoryChanged(new KlipperSystemMemoryChangedEventArgs()
-                {
-                    SystemMemory = value,
-                    SessonId = SessionId,
-                    CallbackId = -1,
-                });
-            }
-        }
-#endif
 
         [ObservableProperty, Obsolete("Use Toolheads instead")]
         [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
@@ -855,16 +681,6 @@ namespace AndreasReitberger.API.Moonraker
             });
         }
 
-        [ObservableProperty, Obsolete("Replaced by the base collection WebCams")]
-        [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-        List<KlipperDatabaseWebcamConfig> webCamConfigs = new();
-        partial void OnWebCamConfigsChanged(List<KlipperDatabaseWebcamConfig> value)
-        {
-            OnKlipperWebCamConfigChanged(new KlipperWebCamConfigChangedEventArgs()
-            {
-                NewConfig = value,
-            });
-        }
 
         [ObservableProperty]
         [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
@@ -876,14 +692,6 @@ namespace AndreasReitberger.API.Moonraker
                 NewPresets = value,
             });
         }
-        #endregion
-
-        #region Database
-
-        [ObservableProperty]
-        [property: JsonIgnore, System.Text.Json.Serialization.JsonIgnore, XmlIgnore]
-        List<string> availableNamespaces = new();
-
         #endregion
 
         #region ReadOnly
@@ -1045,196 +853,7 @@ namespace AndreasReitberger.API.Moonraker
         #region Private
 
         #region RestApi
-        /*
-        async Task<KlipperApiRequestRespone> SendRestApiRequestAsyncOld(
-            MoonrakerCommandBase commandBase,
-            Method method,
-            string command,
-            object jsonObject = null,
-            CancellationTokenSource cts = default,
-            Dictionary<string, string> urlSegments = null,
-            string requestTargetUri = ""
-            )
-        {
-            KlipperApiRequestRespone apiRsponeResult = new() { IsOnline = IsOnline };
-            if (!IsOnline) return apiRsponeResult;
-
-            try
-            {
-                if (cts == default)
-                {
-                    cts = new(DefaultTimeout);
-                }
-                // https://github.com/Arksine/moonraker/blob/master/docs/web_api.md
-                if (restClient == null)
-                {
-                    UpdateRestClientInstance();
-                }
-                RestRequest request = new(
-                    $"{(string.IsNullOrEmpty(requestTargetUri) ? commandBase.ToString() : requestTargetUri)}/{command}")
-                {
-                    RequestFormat = DataFormat.Json,
-                    Method = method
-                };
-
-                bool validHeader = false;
-                // Prefer usertoken over api key
-                if (!string.IsNullOrEmpty(UserToken))
-                {
-                    request.AddHeader("Authorization", $"Bearer {UserToken}");
-                    validHeader = true;
-                }
-                else if (!string.IsNullOrEmpty(ApiKey))
-                {
-                    request.AddHeader("X-Api-Key", $"{ApiKey}");
-                    validHeader = true;
-                }
-
-                if (urlSegments != null)
-                {
-                    foreach (KeyValuePair<string, string> pair in urlSegments)
-                    {
-                        request.AddParameter(pair.Key, pair.Value, ParameterType.QueryString);
-                    }
-                }
-
-                if (jsonObject != null)
-                {
-                    request.AddJsonBody(jsonObject, "application/json");
-                }
-                // https://moonraker.readthedocs.io/en/latest/web_api/#authorization
-                if (!validHeader)
-                {
-                    // Prefer usertoken over api key
-                    if (!string.IsNullOrEmpty(RefreshToken))
-                    {
-                        request.AddParameter("access_token", RefreshToken, ParameterType.QueryString);
-                    }
-                    else if (!string.IsNullOrEmpty(ApiKey))
-                    {
-                        request.AddParameter("token", ApiKey, ParameterType.QueryString);
-                    }
-                }
-
-                Uri fullUri = restClient.BuildUri(request);
-                try
-                {
-                    RestResponse respone = await restClient.ExecuteAsync(request, cts.Token).ConfigureAwait(false);
-                    apiRsponeResult = ValidateRespone(respone, fullUri);
-                }
-                catch (TaskCanceledException texp)
-                {
-                    // Throws exception on timeout, not actually an error but indicates if the server is reachable.
-                    if (!IsOnline)
-                    {
-                        OnError(new UnhandledExceptionEventArgs(texp, false));
-                    }
-                }
-                catch (HttpRequestException hexp)
-                {
-                    // Throws exception on timeout, not actually an error but indicates if the server is not reachable.
-                    if (!IsOnline)
-                    {
-                        OnError(new UnhandledExceptionEventArgs(hexp, false));
-                    }
-                }
-                catch (TimeoutException toexp)
-                {
-                    // Throws exception on timeout, not actually an error but indicates if the server is not reachable.
-                    if (!IsOnline)
-                    {
-                        OnError(new UnhandledExceptionEventArgs(toexp, false));
-                    }
-                }
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-            }
-            return apiRsponeResult;
-        }
-
-        async Task<KlipperApiRequestRespone> SendOnlineCheckRestApiRequestAsyncOld(
-            MoonrakerCommandBase commandBase,
-            string command,
-            CancellationTokenSource cts,
-            string requestTargetUri = ""
-            )
-        {
-            KlipperApiRequestRespone apiRsponeResult = new() { IsOnline = false };
-            try
-            {
-                if (cts == default)
-                {
-                    cts = new(DefaultTimeout);
-                }
-                // https://github.com/Arksine/moonraker/blob/master/docs/web_api.md
-                if (restClient == null)
-                {
-                    UpdateRestClientInstance();
-                }
-                RestRequest request = new(
-                    $"{(string.IsNullOrEmpty(requestTargetUri) ? commandBase.ToString() : requestTargetUri)}/{command}")
-                {
-                    RequestFormat = DataFormat.Json,
-                    Method = Method.Get
-                };
-
-                bool validHeader = false;
-                // Prefer usertoken over api key
-                if (!string.IsNullOrEmpty(UserToken))
-                {
-                    //request.AddHeader("Authorization", $"Bearer {UserToken}", false);
-                    request.AddHeader("Authorization", $"Bearer {UserToken}");
-                    validHeader = true;
-                }
-                else if (!string.IsNullOrEmpty(ApiKey))
-                {
-                    request.AddHeader("X-Api-Key", $"{ApiKey}");
-                    validHeader = true;
-                }
-
-                // https://moonraker.readthedocs.io/en/latest/web_api/#authorization
-                if (!validHeader)
-                {
-                    // Prefer usertoken over api key
-                    if (!string.IsNullOrEmpty(RefreshToken))
-                    {
-                        request.AddParameter("access_token", RefreshToken, ParameterType.QueryString);
-                    }
-                    else if (!string.IsNullOrEmpty(ApiKey))
-                    {
-                        request.AddParameter("token", ApiKey, ParameterType.QueryString);
-                    }
-                }
-
-                Uri fullUri = restClient.BuildUri(request);
-                try
-                {
-                    RestResponse respone = await restClient.ExecuteAsync(request, cts.Token).ConfigureAwait(false);
-                    apiRsponeResult = ValidateRespone(respone, fullUri);
-                }
-                catch (TaskCanceledException)
-                {
-                    // Throws exception on timeout, not actually an error but indicates if the server is not reachable.
-                }
-                catch (HttpRequestException)
-                {
-                    // Throws exception on timeout, not actually an error but indicates if the server is not reachable.
-                }
-                catch (TimeoutException)
-                {
-                    // Throws exception on timeout, not actually an error but indicates if the server is not reachable.
-                }
-
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-            }
-            return apiRsponeResult;
-        }
-        */
+        [Obsolete("Check if this method can be replaced by the base method")]
         async Task<IRestApiRequestRespone> SendMultipartFormDataFileRestApiRequestAsync(
             string filePath,
             string root = "gcodes",
@@ -1326,6 +945,7 @@ namespace AndreasReitberger.API.Moonraker
             return apiRsponeResult;
         }
 
+        [Obsolete("Check if this method can be replaced by the base method")]
         async Task<IRestApiRequestRespone> SendMultipartFormDataFileRestApiRequestAsync(
             string fileName,
             byte[] file,
@@ -1426,50 +1046,12 @@ namespace AndreasReitberger.API.Moonraker
 
         #endregion
 
-        #region Timers
-        [Obsolete("Unused now, will be deleted later")]
-        void StopPingTimer()
-        {
-            if (PingTimer != null)
-            {
-                try
-                {
-                    PingTimer.Change(Timeout.Infinite, Timeout.Infinite);
-                    PingTimer = null;
-                    IsListeningToWebsocket = false;
-                }
-                catch (ObjectDisposedException)
-                {
-                    //PingTimer = null;
-                }
-            }
-        }
-        [Obsolete("Unused now, will be deleted later")]
-        void StopTimer()
-        {
-            if (Timer != null)
-            {
-                try
-                {
-                    Timer.Change(Timeout.Infinite, Timeout.Infinite);
-                    Timer = null;
-                    IsListening = false;
-                }
-                catch (ObjectDisposedException)
-                {
-                    //PingTimer = null;
-                }
-            }
-        }
-        #endregion
-
         #region State & Config
         void UpdateServerConfig(KlipperServerConfig newConfig)
         {
             try
             {
                 if (newConfig == null) return;
-
             }
             catch (Exception exc)
             {
@@ -1692,7 +1274,7 @@ namespace AndreasReitberger.API.Moonraker
 
         #region CheckOnline
 
-        public new async Task CheckOnlineAsync(int timeout = 10000)
+        public override async Task CheckOnlineAsync(int timeout = 10000)
         {
             CancellationTokenSource cts = new(new TimeSpan(0, 0, 0, 0, timeout));
             await CheckOnlineAsync(cts).ConfigureAwait(false);
@@ -1703,64 +1285,6 @@ namespace AndreasReitberger.API.Moonraker
         public Task<bool> CheckIfApiIsValidAsync(int timeout = 10000) => CheckIfApiIsValidAsync($"{MoonrakerCommands.Base}", AuthHeaders, "version", timeout);
         
         public Task CheckServerIfApiIsValidAsync(int timeout = 10000) => CheckIfApiIsValidAsync(timeout);
-        #endregion
-
-        #region WebCam
-        [Obsolete("Replaced by base method")]
-        public string GetDefaultWebCamUriOld()
-        {
-            try
-            {
-                string token = !string.IsNullOrEmpty(ApiKey) ? ApiKey : UserToken;
-                return $"{FullWebAddress}/webcam/?action=stream{(!string.IsNullOrEmpty(token) ? $"?t={token}" : "")}";
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return "";
-            }
-        }
-        [Obsolete("Replaced by base method")]
-        public async Task<string> GetWebCamUriAsyncOld(int index = 0, bool refreshWebCamConfig = false)
-        {
-            try
-            {
-                if (WebCamConfigs?.Count <= 0 || refreshWebCamConfig)
-                {
-                    await RefreshWebCamConfigAsync().ConfigureAwait(false);
-                }
-                KlipperDatabaseWebcamConfig config = null;
-                if (WebCamConfigs?.Count > index)
-                {
-                    config = WebCamConfigs[index];
-                }
-                else if (WebCamConfigs.Count > 0)
-                {
-                    config = WebCamConfigs.FirstOrDefault();
-                }
-                // If nothing is found, try the default setup
-                else
-                {
-                    config = new KlipperDatabaseWebcamConfig()
-                    {
-                        Id = Guid.NewGuid(),
-                        Alias = "Default",
-                        Enabled = true,
-                        FlipX = false,
-                        FlipY = false,
-                        TargetFps = 15,
-                        Url = "/webcam?action=stream"
-                    };
-                }
-                string token = !string.IsNullOrEmpty(ApiKey) ? ApiKey : UserToken;
-                return config == null ? GetDefaultWebCamUri() : $"{FullWebAddress}{config.Url}{(!string.IsNullOrEmpty(token) ? $"&t={token}" : "")}";
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return "";
-            }
-        }
         #endregion
 
         #region Updates
@@ -1886,1203 +1410,6 @@ namespace AndreasReitberger.API.Moonraker
             }
         }
 
-        #endregion
-
-        #region Printer Administration
-
-        public async Task RefreshPrinterInfoAsync()
-        {
-            try
-            {
-                KlipperPrinterStateMessageResult result = await GetPrinterInfoAsync().ConfigureAwait(false);
-                PrinterInfo = result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                PrinterInfo = null;
-            }
-        }
-
-        public async Task<KlipperPrinterStateMessageResult> GetPrinterInfoAsync()
-        {
-            IRestApiRequestRespone result = null;
-            KlipperPrinterStateMessageResult resultObject = null;
-            try
-            {
-                //object cmd = new { name = ScriptName };
-
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "info",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegements,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Get, "info")
-                    .ConfigureAwait(false);
-                */
-                KlipperPrinterStateMessageRespone state = GetObjectFromJson<KlipperPrinterStateMessageRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return state?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperPrinterStateMessageRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<bool> EmergencyStopPrinterAsync()
-        {
-            try
-            {
-                //object cmd = new { name = ScriptName };
-
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "emergency_stop",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegements,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                KlipperApiRequestRespone result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Post, "emergency_stop")
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> RestartPrinterAsync()
-        {
-            try
-            {
-                //object cmd = new { name = ScriptName };
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "restart",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegements,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                KlipperApiRequestRespone result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Post, "restart")
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> RestartFirmwareAsync()
-        {
-            try
-            {
-                //object cmd = new { name = ScriptName };
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                IRestApiRequestRespone result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "firmware_restart",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegements,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                KlipperApiRequestRespone result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Post, "firmware_restart")
-                    .ConfigureAwait(false);
-                */
-                return GetQueryResult(result.Result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> SetHeaterBedTargetAsync(int target)
-        {
-            try
-            {
-                string cmd = $"SET_HEATER_TEMPERATURE HEATER=heater_bed TARGET={target}";
-                bool result = await RunGcodeScriptAsync(cmd).ConfigureAwait(false);
-                return result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> SetExtruderTargetAsync(int target, int extruder = 0)
-        {
-            try
-            {
-                string cmd = $"SET_HEATER_TEMPERATURE HEATER=extruder{(extruder <= 0 ? "" : $"{extruder}")} TARGET={target}";
-                bool result = await RunGcodeScriptAsync(cmd).ConfigureAwait(false);
-                return result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> SetSpeedFactorAsync(int target)
-        {
-            try
-            {
-                string cmd = $"M220 S{target}";
-                bool result = await RunGcodeScriptAsync(cmd).ConfigureAwait(false);
-                return result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> SetExtrusionFactorAsync(int target)
-        {
-            try
-            {
-                string cmd = $"M221 S{target}";
-                bool result = await RunGcodeScriptAsync(cmd).ConfigureAwait(false);
-                return result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-
-        public async Task<bool> SetFanSpeedTargetAsync(int target, bool isPercentage, int fanId = 0)
-        {
-            try
-            {
-                int setSpeed = target;
-                if (!isPercentage)
-                {
-                    // Avoid invalid ranges
-                    switch (target)
-                    {
-                        case > 255:
-                            setSpeed = 255;
-                            break;
-                        case < 0:
-                            setSpeed = 0;
-                            break;
-                        default:
-                            break;
-                    }
-                }
-                else
-                {
-                    setSpeed = Convert.ToInt32(target * 255f / 100f);
-                }
-
-                string cmd = $"M106 S{setSpeed}";
-                bool result = await RunGcodeScriptAsync(cmd).ConfigureAwait(false);
-                return result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return false;
-            }
-        }
-        #endregion
-
-        #region Printer Status
-        public async Task<List<string>> GetPrinterObjectListAsync(string startsWith = "", bool removeStartTag = false)
-        {
-            IRestApiRequestRespone result = null;
-            List<string> resultObject = new();
-            try
-            {
-                //object cmd = new { name = ScriptName };
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "objects/list",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegements,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Get, "objects/list")
-                    .ConfigureAwait(false);
-                */
-                KlipperActionListRespone state = GetObjectFromJson<KlipperActionListRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                if (!string.IsNullOrEmpty(startsWith))
-                {
-                    resultObject = state?.Result?.Objects.Where(obj => obj.StartsWith(startsWith)).ToList();
-                    if (removeStartTag)
-                    {
-                        resultObject = resultObject.Select(item => item.Replace(startsWith, string.Empty).Trim()).ToList();
-                    }
-                    return resultObject;
-                }
-                return state?.Result?.Objects;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperActionListRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<Dictionary<string, object>> QueryPrinterObjectStatusAsync(Dictionary<string, string> objects)
-        {
-            IRestApiRequestRespone result = null;
-            Dictionary<string, object> resultObject = new();
-            try
-            {
-                Dictionary<string, string> urlSegments = new();
-                foreach (KeyValuePair<string, string> obj in objects)
-                {
-                    // Do not query macros here, there is an extra method for this.
-                    if (obj.Key.StartsWith("gcode_macro")) continue;
-                    urlSegments.Add(obj.Key, obj.Value);
-                }
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "objects/query",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Get, "objects/query", jsonObject: null, cts: default, urlSegments: urlSegments)
-                    .ConfigureAwait(false);
-                */
-                KlipperPrinterStatusRespone queryResult = GetObjectFromJson<KlipperPrinterStatusRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                if (queryResult?.Result?.Status is JObject jsonObject)
-                {
-                    foreach (JProperty property in jsonObject.Children<JProperty>())
-                    {
-                        Stack<JToken> avilableProperties = new(jsonObject.Children<JToken>());
-                        do
-                        {
-                            JToken token = avilableProperties.Pop();
-                            if (token is JProperty propTest)
-                            {
-                                // Get the childs for this tags
-                                if (propTest.Name.StartsWith("configfile") || propTest.Name.StartsWith("settings"))
-                                {
-                                    // Add all child properties back to the stack
-                                    List<JToken> children = token.Children().ToList();
-                                    foreach (JToken child in children)
-                                    {
-                                        avilableProperties.Push(child);
-                                    }
-                                    continue;
-                                }
-                            }
-                            else if (token is JToken childToken)
-                            {
-                                if (childToken?.First is not JProperty jp)
-                                    continue;
-                                /**/
-                                // Get the childs for this tags
-                                if (jp.Name.StartsWith("configfile") || jp.Name.StartsWith("settings"))
-                                {
-                                    // Add all child properties back to the stack
-                                    List<JToken> children = token.Children().ToList();
-                                    foreach (JToken child in children)
-                                    {
-                                        avilableProperties.Push(child);
-                                    }
-                                    continue;
-                                }
-
-                            }
-
-                            if (token is not JProperty parent)
-                            {
-                                // Add all child properties back to the stack
-                                List<JToken> chilTokens = token.Children().ToList();
-                                foreach (JToken child in chilTokens)
-                                {
-                                    avilableProperties.Push(child);
-                                }
-                                continue;
-                            }
-
-                            string name = parent.Name;
-                            string path = parent.Path;
-                            string jsonBody = parent.Value.ToString();
-                            switch (name)
-                            {
-                                case "probe":
-                                    KlipperStatusProbe probe =
-                                        GetObjectFromJson<KlipperStatusProbe>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, probe);
-                                    break;
-                                case "configfile":
-                                    KlipperStatusConfigfile configFile =
-                                        GetObjectFromJson<KlipperStatusConfigfile>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, configFile);
-                                    break;
-                                case "query_endstops":
-                                    KlipperStatusQueryEndstops endstops =
-                                        GetObjectFromJson<KlipperStatusQueryEndstops>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, endstops);
-                                    break;
-                                case "virtual_sdcard":
-                                    KlipperStatusVirtualSdcard virtualSdcardState =
-                                        GetObjectFromJson<KlipperStatusVirtualSdcard>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, virtualSdcardState);
-                                    break;
-                                case "display_status":
-                                    KlipperStatusDisplay displayState =
-                                        GetObjectFromJson<KlipperStatusDisplay>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, displayState);
-                                    break;
-                                case "moonraker_stats":
-                                    MoonrakerStatInfo notifyProcState =
-                                        GetObjectFromJson<MoonrakerStatInfo>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, notifyProcState);
-                                    break;
-                                case "mcu":
-                                    KlipperStatusMcu mcuState =
-                                        GetObjectFromJson<KlipperStatusMcu>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, mcuState);
-                                    break;
-                                case "system_stats":
-                                    KlipperStatusSystemStats systemState =
-                                        GetObjectFromJson<KlipperStatusSystemStats>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, systemState);
-                                    break;
-                                case "cpu_temp":
-                                    double cpuTemp =
-                                        GetObjectFromJson<double>(jsonBody.Replace(",", "."));
-                                    resultObject.Add(name, cpuTemp);
-                                    break;
-                                case "websocket_connections":
-                                    int wsConnections =
-                                        GetObjectFromJson<int>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, wsConnections);
-                                    break;
-                                case "network":
-                                    Dictionary<string, KlipperNetworkInterface> network =
-                                        GetObjectFromJson<Dictionary<string, KlipperNetworkInterface>>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, network);
-                                    break;
-                                case "gcode_move":
-                                    KlipperStatusGcodeMove gcodeMoveState =
-                                        GetObjectFromJson<KlipperStatusGcodeMove>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, gcodeMoveState);
-                                    break;
-                                case "print_stats":
-                                    KlipperStatusPrintStats printStats =
-                                        GetObjectFromJson<KlipperStatusPrintStats>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    printStats.ValidPrintState = jsonBody.Contains("state");
-                                    resultObject.Add(name, printStats);
-                                    break;
-                                case "fan":
-                                    KlipperStatusFan fanState =
-                                        GetObjectFromJson<KlipperStatusFan>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, fanState);
-                                    break;
-                                case "toolhead":
-                                    KlipperStatusToolhead toolhead =
-                                        GetObjectFromJson<KlipperStatusToolhead>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, toolhead);
-                                    break;
-                                case "heater_bed":
-                                    // In the status report the temp is missing, so do not parse the heater then.
-                                    //if (!jsonBody.Contains("temperature")) break;
-                                    if (path.EndsWith("settings.heater_bed"))
-                                    {
-                                        KlipperConfigHeaterBed settingsHeaterBed =
-                                            GetObjectFromJson<KlipperConfigHeaterBed>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                        resultObject.Add(name, settingsHeaterBed);
-                                    }
-                                    else
-                                    {
-                                        KlipperStatusHeaterBed heaterBed =
-                                            GetObjectFromJson<KlipperStatusHeaterBed>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                        resultObject.Add(name, heaterBed);
-                                    }
-                                    break;
-                                case "extruder":
-                                case "extruder1":
-                                case "extruder2":
-                                case "extruder3":
-                                    // In the status report the temp is missing, so do not parse the heater then.
-                                    //if (!jsonBody.Contains("temperature")) break;
-                                    if (path.EndsWith("settings.extruder"))
-                                    {
-                                        KlipperConfigExtruder settingsExtruder =
-                                            GetObjectFromJson<KlipperConfigExtruder>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                        resultObject.Add(name, settingsExtruder);
-                                    }
-                                    else
-                                    {
-                                        KlipperStatusExtruder extruder =
-                                            GetObjectFromJson<KlipperStatusExtruder>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                        resultObject.Add(name, extruder);
-                                    }
-                                    break;
-                                case "motion_report":
-                                    KlipperStatusMotionReport motionReport =
-                                        GetObjectFromJson<KlipperStatusMotionReport>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, motionReport);
-                                    break;
-                                case "idle_timeout":
-                                    KlipperStatusIdleTimeout idleTimeout =
-                                        GetObjectFromJson<KlipperStatusIdleTimeout>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    idleTimeout.ValidState = jsonBody.Contains("state");
-                                    resultObject.Add(name, idleTimeout);
-                                    break;
-                                case "filament_switch_sensor fsensor":
-                                    KlipperStatusFilamentSensor fSensor =
-                                        GetObjectFromJson<KlipperStatusFilamentSensor>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, fSensor);
-                                    break;
-                                case "pause_resume":
-                                    KlipperStatusPauseResume pauseResume =
-                                        GetObjectFromJson<KlipperStatusPauseResume>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, pauseResume);
-                                    break;
-                                case "action":
-                                    string action = jsonBody;
-                                    resultObject.Add(name, action);
-                                    break;
-                                case "bed_mesh":
-                                    KlipperStatusMesh mesh =
-                                        GetObjectFromJson<KlipperStatusMesh>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, mesh);
-                                    break;
-                                case "job":
-                                    KlipperStatusJob job =
-                                        GetObjectFromJson<KlipperStatusJob>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                    resultObject.Add(name, job);
-                                    break;
-                                default:
-#if DEBUG
-                                    Console.WriteLine($"No Json object found for '{name}' => '{jsonBody}");
-#endif
-                                    if (name.StartsWith("gcode_macro"))
-                                    {
-                                        KlipperGcodeMacro gcMacro =
-                                            GetObjectFromJson<KlipperGcodeMacro>(jsonBody, NewtonsoftJsonSerializerSettings);
-                                        if (string.IsNullOrEmpty(gcMacro?.Name))
-                                        {
-                                            gcMacro.Name = name.Replace("gcode_macro", string.Empty).Trim();
-                                        }
-                                        resultObject.Add(name, gcMacro);
-                                    }
-                                    else
-                                    {
-                                        // If no parser found, pass the json object instead
-                                        resultObject.Add(name, parent.Value);
-                                    }
-#if ConcurrentDictionary
-                                    ConcurrentDictionary<string, string> loggedResults = new(IgnoredJsonResults);
-#else
-                                    Dictionary<string, string> loggedResults = new(IgnoredJsonResults);
-#endif
-                                    if (!loggedResults.ContainsKey(name))
-                                    {
-                                        // Log unused json results for further releases
-#if ConcurrentDictionary
-                                        loggedResults.TryAdd(name, jsonBody);
-#else
-                                        loggedResults.Add(name, jsonBody);
-#endif
-                                        IgnoredJsonResults = loggedResults;
-                                    }
-                                    break;
-                            }
-
-                        }
-                        while (avilableProperties.Count > 0);
-                    }
-                }
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<Dictionary<string, KlipperGcodeMacro>> GetGcodeMacrosAsync()
-        {
-            IRestApiRequestRespone result = null;
-            Dictionary<string, KlipperGcodeMacro> resultObject = new();
-            try
-            {
-                Dictionary<string, string> objects = new()
-                {
-                    { "configfile", "settings" }
-                };
-
-                Dictionary<string, object> settings = await QueryPrinterObjectStatusAsync(objects).ConfigureAwait(false);
-#if NETSTANDARD || NET6_0_OR_GREATER
-                IEnumerable<KeyValuePair<string, KlipperGcodeMacro>> macros =
-                    settings.Where(keypair => keypair.Key.StartsWith("gcode_macro"))
-                    .Select(pair => new KeyValuePair<string, KlipperGcodeMacro>(pair.Key, pair.Value as KlipperGcodeMacro));
-                return new(macros);
-#else
-                List<KeyValuePair<string, KlipperGcodeMacro>> macros =
-                    settings.Where(keypair => keypair.Key.StartsWith("gcode_macro"))
-                    .Select(pair => new KeyValuePair<string, KlipperGcodeMacro>(pair.Key, pair.Value as KlipperGcodeMacro))
-                .ToList()
-                ;
-
-                resultObject = new();
-                for (int i = 0; i < macros?.Count; i++)
-                {
-                    resultObject.Add(macros[i].Key, macros[i].Value);
-                }
-                return resultObject;
-#endif
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task<Dictionary<string, KlipperStatusFilamentSensor>> GetFilamentSensorsAsync(Dictionary<string, string> macros = null)
-        {
-            IRestApiRequestRespone result = null;
-            Dictionary<string, KlipperStatusFilamentSensor> resultObject = new();
-            try
-            {
-                Dictionary<string, string> urlSegments = new();
-                if (macros != null)
-                {
-                    foreach (KeyValuePair<string, string> obj in macros)
-                    {
-                        urlSegments.Add(obj.Key, obj.Value);
-                    }
-                }
-                else
-                {
-                    urlSegments.Add("filament_switch_sensor", string.Empty);
-                }
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "objects/query",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Get, "objects/query", jsonObject: null, cts: default, urlSegments: urlSegments)
-                    .ConfigureAwait(false);
-                */
-                KlipperFilamentSensorsRespone queryResult = GetObjectFromJson<KlipperFilamentSensorsRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result?.Status;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperFilamentSensorsRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<KlipperStatusPrintStats> GetPrintStatusAsync()
-        {
-            KlipperStatusPrintStats resultObject = null;
-            try
-            {
-                // Doc: https://moonraker.readthedocs.io/en/latest/printer_objects/#print_stats
-                string key = "print_stats";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, "" }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusPrintStats stateObj)
-                {
-                    stateObj.ValidPrintState = true;
-                    //IsPrinting = stateObj.State == KlipperPrintStates.Printing;
-                    resultObject = stateObj;
-                }
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshPrintStatusAsync()
-        {
-            try
-            {
-                KlipperStatusPrintStats result = await GetPrintStatusAsync().ConfigureAwait(false);
-                PrintStats = result;
-                if (PrintStats != null)
-                {
-                    await RefreshGcodeMetadataAsync(PrintStats.Filename).ConfigureAwait(false);
-                }
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                PrintStats = null;
-                GcodeMeta = null;
-            }
-        }
-
-        public async Task<KlipperStatusExtruder> GetExtruderStatusAsync(int index = 0)
-        {
-            KlipperStatusExtruder resultObject = null;
-            try
-            {
-                string key = $"extruder{(index > 0 ? index : "")}";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, "" }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusExtruder stateObj)
-                    resultObject = stateObj;
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshExtruderStatusAsync(int index = 0)
-        {
-            try
-            {
-                KlipperStatusExtruder result = await GetExtruderStatusAsync(index).ConfigureAwait(false);
-                if (result != null)
-                {
-#if ConcurrentDictionary
-                    ConcurrentDictionary<int, IToolhead> states = new(Toolheads);
-                    states.AddOrUpdate(index, result, (key, oldValue) => oldValue = result);
-#else
-                    Dictionary<int, KlipperStatusExtruder> states = new()
-                    {
-                        { index, result }
-                    };
-#endif
-                    //Extruders = states;
-                    Toolheads = states;
-                }
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Toolheads = new();
-            }
-        }
-
-        public async Task<KlipperStatusFan> GetFanStatusAsync()
-        {
-            KlipperStatusFan resultObject = null;
-            try
-            {
-                string key = "fan";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, "" }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusFan stateObj)
-                    resultObject = stateObj;
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshFanStatusAsync()
-        {
-            try
-            {
-                KlipperStatusFan result = await GetFanStatusAsync().ConfigureAwait(false);
-                Fan = result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Fan = null;
-            }
-        }
-
-        public async Task<KlipperStatusIdleTimeout> GetIdleStatusAsync()
-        {
-            KlipperStatusIdleTimeout resultObject = null;
-            try
-            {
-                string key = "idle_timeout";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, string.Empty }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusIdleTimeout stateObj)
-                {
-                    stateObj.ValidState = true;
-                    resultObject = stateObj;
-                }
-                return resultObject;
-
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshIdleStatusAsync()
-        {
-            try
-            {
-                KlipperStatusIdleTimeout result = await GetIdleStatusAsync().ConfigureAwait(false);
-                IdleState = result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                IdleState = null;
-            }
-        }
-
-        public async Task<KlipperStatusDisplay> GetDisplayStatusAsync()
-        {
-            KlipperStatusDisplay resultObject = null;
-            try
-            {
-                string key = "display_status";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, string.Empty }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusDisplay stateObj)
-                    resultObject = stateObj;
-                return resultObject;
-
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshDisplayStatusAsync()
-        {
-            try
-            {
-                KlipperStatusDisplay result = await GetDisplayStatusAsync().ConfigureAwait(false);
-                DisplayStatus = result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                DisplayStatus = null;
-            }
-        }
-
-        public async Task<KlipperStatusToolhead> GetToolHeadStatusAsync()
-        {
-            KlipperStatusToolhead resultObject = null;
-            try
-            {
-                string key = "toolhead";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, string.Empty }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusToolhead stateObj)
-                    resultObject = stateObj;
-                return resultObject;
-
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshToolHeadStatusAsync()
-        {
-            try
-            {
-                KlipperStatusToolhead result = await GetToolHeadStatusAsync().ConfigureAwait(false);
-                ToolHead = result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                ToolHead = null;
-            }
-        }
-
-        public async Task<KlipperStatusGcodeMove> GetGcodeMoveStatusAsync()
-        {
-            KlipperStatusGcodeMove resultObject = null;
-            try
-            {
-                string key = "gcode_move";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, string.Empty }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusGcodeMove stateObj)
-                    resultObject = stateObj;
-                return resultObject;
-
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshGcodeMoveStatusAsync()
-        {
-            try
-            {
-                KlipperStatusGcodeMove result = await GetGcodeMoveStatusAsync().ConfigureAwait(false);
-                GcodeMove = result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                GcodeMove = null;
-            }
-        }
-
-        public async Task<KlipperStatusMotionReport> GetMotionReportAsync()
-        {
-            KlipperStatusMotionReport resultObject = null;
-            try
-            {
-                string key = "motion_report";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, string.Empty }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusMotionReport stateObj)
-                    resultObject = stateObj;
-                return resultObject;
-
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshMotionReportAsync()
-        {
-            try
-            {
-                KlipperStatusMotionReport result = await GetMotionReportAsync().ConfigureAwait(false);
-                MotionReport = result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                MotionReport = null;
-            }
-        }
-
-        public async Task<KlipperStatusVirtualSdcard> GetVirtualSdCardStatusAsync()
-        {
-            KlipperStatusVirtualSdcard resultObject = null;
-            try
-            {
-                string key = "virtual_sdcard";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, string.Empty }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusVirtualSdcard stateObj)
-                    resultObject = stateObj;
-                return resultObject;
-
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshVirtualSdCardStatusAsync()
-        {
-            try
-            {
-                KlipperStatusVirtualSdcard result = await GetVirtualSdCardStatusAsync().ConfigureAwait(false);
-                VirtualSdCard = result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                VirtualSdCard = null;
-            }
-        }
-
-        public async Task<KlipperStatusHeaterBed> GetHeaterBedStatusAsync()
-        {
-            KlipperStatusHeaterBed resultObject = null;
-            try
-            {
-                string key = "heater_bed";
-                Dictionary<string, string> queryObjects = new()
-                {
-                    { key, string.Empty }
-                };
-
-                Dictionary<string, object> result = await QueryPrinterObjectStatusAsync(queryObjects)
-                    .ConfigureAwait(false);
-
-                if (result.ContainsKey(key) && result?[key] is KlipperStatusHeaterBed stateObj)
-                {
-                    resultObject = stateObj;
-                }
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshHeaterBedStatusAsync()
-        {
-            try
-            {
-                KlipperStatusHeaterBed result = await GetHeaterBedStatusAsync().ConfigureAwait(false);
-                ConcurrentDictionary<int, IHeaterComponent> heaters = new(HeatedBeds);
-                heaters.AddOrUpdate(0, result, (key, oldValue) => oldValue = result);
-                HeatedBeds = heaters;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                HeatedBeds = new();
-            }
-        }
-
-        public Task<string> SubscribePrinterObjectStatusAsync(long? connectionId, List<string> objects) => SubscribePrinterObjectStatusAsync((long)connectionId, objects);
-        public async Task<string> SubscribePrinterObjectStatusAsync(long connectionId, List<string> objects)
-        {
-            IRestApiRequestRespone result = null;
-            string resultObject = null;
-            try
-            {
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "connection_id", $"{connectionId}" }
-                };
-
-                for (int i = 0; i < objects.Count; i++)
-                {
-                    string key = objects[i];
-                    string value = string.Empty;
-                    urlSegments.Add(key, value);
-                }
-
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "objects/subscribe",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Post, "objects/subscribe", jsonObject: null, cts: default, urlSegments: urlSegments)
-                    .ConfigureAwait(false);
-                */
-                return result?.Result;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<string> SubscribeAllPrinterObjectStatusAsync(long? connectionId)
-        {
-            List<string> objects = await GetPrinterObjectListAsync().ConfigureAwait(false);
-            return await SubscribePrinterObjectStatusAsync((long)connectionId, objects).ConfigureAwait(false);
-        }
-        public async Task<KlipperEndstopQueryResult> QueryEndstopsAsync()
-        {
-            IRestApiRequestRespone result = null;
-            KlipperEndstopQueryResult resultObject = null;
-            try
-            {
-                string targetUri = $"{MoonrakerCommands.Printer}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "query_endstops/status",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegements,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                //result = await SendRestApiRequestAsync(MoonrakerCommandBase.printer, Method.Get, "query_endstops/status").ConfigureAwait(false);
-                KlipperEndstopQueryRespone queryResult = GetObjectFromJson<KlipperEndstopQueryRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperEndstopQueryRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
         #endregion
 
         #region Server Config
@@ -3277,40 +1604,6 @@ namespace AndreasReitberger.API.Moonraker
                 return false;
             }
         }
-        #endregion
-
-        #region WebSocket   
-
-        /* Not available for HTTP
-        public async Task<KlipperAccessTokenResult> GetWebSocketIdAsync()
-        {
-            IRestApiRequestRespone result = null;
-            KlipperAccessTokenResult resultObject = null;
-            try
-            {
-                //object cmd = new { name = ScriptName };
-                result = await SendRestApiRequestAsync(MoonRakerCommandBase.server, Method.Get, "websocket_id").ConfigureAwait(false);
-                KlipperAccessTokenResult accessToken = GetObjectFromJson<KlipperAccessTokenResult>(result.Result, NewtonsoftJsonSerializerSettings);
-                SessionId = accessToken?.Result;
-                return accessToken;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        */
         #endregion
 
         #region Gcode API
@@ -3867,740 +2160,7 @@ namespace AndreasReitberger.API.Moonraker
             }
         }
         #endregion
-
-        #region File Operations
-        // Doc: https://github.com/Arksine/moonraker/blob/master/docs/web_api.md#list-available-files
-
-        public async Task RefreshAvailableFilesAsync(string rootPath = "", bool includeGcodeMeta = true)
-        {
-            try
-            {
-                Files = await GetAvailableFilesAsync(rootPath, includeGcodeMeta).ConfigureAwait(false);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Files = new ObservableCollection<IGcode>();
-            }
-        }
-
-        public async Task<ObservableCollection<IGcode>> GetAvailableFilesAsync(string rootPath = "", bool includeGcodeMeta = true)
-        {
-            IRestApiRequestRespone result = null;
-            ObservableCollection<IGcode> resultObject = new();
-            try
-            {
-                Dictionary<string, string> urlSegments = new();
-                if (!string.IsNullOrEmpty(rootPath))
-                {
-                    urlSegments.Add("root", rootPath);
-                }
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "files/list",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                //result = await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Get, "files/list", default, null, urlSegments).ConfigureAwait(false);
-                KlipperFileListRespone files = GetObjectFromJson<KlipperFileListRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                if (includeGcodeMeta)
-                {
-                    for (int i = 0; i < files?.Result?.Count; i++)
-                    {
-                        KlipperFile current = files?.Result[i];
-                        current.Meta = await GetGcodeMetadataAsync(current.FilePath).ConfigureAwait(false);
-                        if (current.Meta?.GcodeImages?.Count > 0)
-                        {
-                            current.Image = await GetGcodeSecondThumbnailImageAsync(current?.Meta)
-                                .ConfigureAwait(false)
-                                ;
-                        }
-                    }
-                }
-                return new(files?.Result ?? new());
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperFileListRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task<List<IGcode>> GetAvailableFilesAsListAsync(string rootPath = "")
-        {
-            List<IGcode> resultObject = new();
-            try
-            {
-                ObservableCollection<IGcode> result = await GetAvailableFilesAsync(rootPath).ConfigureAwait(false);
-                return result?.ToList();
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public new Task<ObservableCollection<IGcode>> GetFilesAsync() => GetAvailableFilesAsync();
-        
-        public async Task<KlipperGcodeMetaResult> GetGcodeMetadataAsync(string fileName)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperGcodeMetaResult resultObject = null;
-            try
-            {
-                if (string.IsNullOrEmpty(fileName)) return resultObject;
-
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "filename", fileName }
-                };
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "files/metadata",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                //result = await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Get, "files/metadata", default, null, urlSegments).ConfigureAwait(false);
-                KlipperGcodeMetaRespone queryResult = GetObjectFromJson<KlipperGcodeMetaRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperGcodeMetaRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshGcodeMetadataAsync(string fileName)
-        {
-            try
-            {
-                KlipperGcodeMetaResult meta = await GetGcodeMetadataAsync(fileName).ConfigureAwait(false);
-                GcodeMeta = meta;
-                if (PrintStats?.State == KlipperPrintStates.Printing)
-                {
-                    // Get current print image 
-                    CurrentPrintImage = await GetGcodeLargestThumbnailImageAsync(GcodeMeta);
-                }
-                else
-                {
-                    CurrentPrintImage = Array.Empty<byte>();
-                }
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                GcodeMeta = null;
-            }
-        }
-
-        public async Task<byte[]> GetGcodeThumbnailImageAsync(string relativePath, int timeout = 10000)
-        {
-            try
-            {
-                //Uri target = new($"{FullWebAddress}/server/files/gcodes/{relativePath}");
-                string target = $"{FullWebAddress}/server/files/gcodes/{relativePath}";
-                byte[] thumb = await DownloadFileFromUriAsync(target, timeout)
-                    .ConfigureAwait(false)
-                    ;
-
-                return thumb;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return Array.Empty<byte>();
-            }
-        }
-        public async Task<byte[]> GetGcodeThumbnailImageAsync(IGcodeMeta gcodeMeta, int index = 0, int timeout = 10000)
-        {
-            if (gcodeMeta is null || gcodeMeta?.GcodeImages is null) return Array.Empty<byte>();
-            string path = gcodeMeta.GcodeImages.Count > index ?
-                gcodeMeta.GcodeImages[index]?.Path : gcodeMeta.GcodeImages.FirstOrDefault()?.Path;
-
-            string subfolder = string.Empty;
-            if (gcodeMeta?.FileName?.Contains("/") ?? false)
-            {
-                subfolder = gcodeMeta.FileName[..gcodeMeta.FileName.LastIndexOf("/")];
-                subfolder += "/";
-            }
-
-            return string.IsNullOrEmpty(path) ? null : await GetGcodeThumbnailImageAsync(subfolder + path, timeout)
-                .ConfigureAwait(false)
-                ;
-        }
-        public async Task<byte[]> GetGcodeLargestThumbnailImageAsync(KlipperGcodeMetaResult gcodeMeta, int timeout = 10000)
-        {
-            if (gcodeMeta is null || gcodeMeta?.GcodeImages is null) return Array.Empty<byte>();
-            string path = gcodeMeta.GcodeImages
-                .OrderByDescending(image => image.Size)
-                .FirstOrDefault()?.Path
-                ;
-
-            string subfolder = string.Empty;
-            if (gcodeMeta?.FileName?.Contains("/") ?? false)
-            {
-                subfolder = gcodeMeta.FileName[..gcodeMeta.FileName.LastIndexOf("/")];
-                subfolder += "/";
-            }
-
-            return string.IsNullOrEmpty(path) ? null : await GetGcodeThumbnailImageAsync(subfolder + path, timeout)
-                .ConfigureAwait(false)
-                ;
-        }
-        public async Task<byte[]> GetGcodeSmallestThumbnailImageAsync(KlipperGcodeMetaResult gcodeMeta, int timeout = 10000)
-        {
-            if (gcodeMeta is null || gcodeMeta?.GcodeImages is null) return Array.Empty<byte>();
-            string path = gcodeMeta.GcodeImages
-                .OrderBy(image => image.Size)
-                .FirstOrDefault()?.Path
-                ;
-
-            string subfolder = string.Empty;
-            if (gcodeMeta?.FileName?.Contains("/") ?? false)
-            {
-                //subfolder = gcodeMeta.Filename.Substring(0, gcodeMeta.Filename.LastIndexOf("/"));
-                subfolder = gcodeMeta.FileName[..gcodeMeta.FileName.LastIndexOf("/")];
-                subfolder += "/";
-            }
-
-            return string.IsNullOrEmpty(path) ? null : await GetGcodeThumbnailImageAsync(subfolder + path, timeout)
-                .ConfigureAwait(false)
-                ;
-        }
-        public async Task<byte[]> GetGcodeSecondThumbnailImageAsync(IGcodeMeta gcodeMeta, int timeout = 10000)
-        {
-            if (gcodeMeta is null || gcodeMeta?.GcodeImages is null) return Array.Empty<byte>();
-            string path = gcodeMeta.GcodeImages
-                .OrderBy(image => image.Size)?
-                .Skip(1)? // Skipped the smallest image
-                .FirstOrDefault()?.Path
-                ;
-
-            string subfolder = string.Empty;
-            if (gcodeMeta?.FileName?.Contains("/") ?? false)
-            {
-                subfolder = gcodeMeta.FileName[..gcodeMeta.FileName.LastIndexOf("/")];
-                subfolder += "/";
-            }
-
-            return string.IsNullOrEmpty(path) ? null : await GetGcodeThumbnailImageAsync(subfolder + path, timeout)
-                .ConfigureAwait(false)
-                ;
-        }
-        public async Task RefreshDirectoryInformationAsync(string path = "", bool extended = true)
-        {
-            try
-            {
-                if (string.IsNullOrEmpty(path))
-                    path = "gcodes";
-                KlipperDirectoryInfoResult result = await GetDirectoryInformationAsync(path, extended).ConfigureAwait(false);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-            }
-        }
-
-        public async Task<KlipperDirectoryInfoResult> GetDirectoryInformationAsync(string path, bool extended = true)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperDirectoryInfoResult resultObject = null;
-            try
-            {
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "path", path },
-                    { "extended", extended ? "true" : "false" }
-                };
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "files/directory",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                //result = await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Get, "files/directory", default, null, urlSegments).ConfigureAwait(false);
-
-                KlipperDirectoryInfoRespone queryResult = GetObjectFromJson<KlipperDirectoryInfoRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                if (queryResult?.Result?.DiskUsage != null)
-                {
-                    FreeDiskSpace = queryResult.Result.DiskUsage.Free;
-                    TotalDiskSpace = queryResult.Result.DiskUsage.Total;
-                    UsedDiskSpace = queryResult.Result.DiskUsage.Used;
-                }
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDirectoryInfoRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<List<KlipperDirectory>> GetAvailableDirectoriesAsync(string path = "")
-        {
-            List<KlipperDirectory> resultObject = new();
-            try
-            {
-                if (string.IsNullOrEmpty(path))
-                {
-                    path = "gcodes";
-                }
-                KlipperDirectoryInfoResult result = await GetDirectoryInformationAsync(path, false).ConfigureAwait(false);
-                resultObject = result?.Dirs ?? new();
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task RefreshAvailableDirectorienAsync(string path = "")
-        {
-            try
-            {
-                if (string.IsNullOrEmpty(path))
-                {
-                    path = "gcodes";
-                }
-                List<KlipperDirectory> result = await GetAvailableDirectoriesAsync(path).ConfigureAwait(false);
-                AvailableDirectories = new(result);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-            }
-        }
-
-        public async Task<KlipperDirectoryActionResult> CreateDirectoryAsync(string directory)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperDirectoryActionResult resultObject = null;
-            try
-            {
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "path", directory }
-                };
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "files/directory",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Post, $"files/directory", jsonObject: null, cts: default, urlSegments: urlSegments)
-                    .ConfigureAwait(false);
-                */
-                KlipperDirectoryActionRespone queryResult = GetObjectFromJson<KlipperDirectoryActionRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDirectoryActionRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<KlipperDirectoryActionResult> DeleteDirectoryAsync(string directory, bool force = false)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperDirectoryActionResult resultObject = null;
-            try
-            {
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "path", directory },
-                    { "force", force ? "true" : "false" }
-                };
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Delete,
-                       command: "files/directory",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Delete, $"files/directory", jsonObject: null, cts: default, urlSegments: urlSegments)
-                    .ConfigureAwait(false);
-                */
-                KlipperDirectoryActionRespone queryResult = GetObjectFromJson<KlipperDirectoryActionRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDirectoryActionRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<KlipperDirectoryActionResult> MoveDirectoryOrFileAsync(string source, string destination)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperDirectoryActionResult resultObject = null;
-            try
-            {
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "source", source },
-                    { "dest", destination }
-                };
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "files/move",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Post, $"files/move", jsonObject: null, cts: default, urlSegments: urlSegments)
-                    .ConfigureAwait(false);
-                */
-                KlipperDirectoryActionRespone queryResult = GetObjectFromJson<KlipperDirectoryActionRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDirectoryActionRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<KlipperDirectoryActionResult> CopyDirectoryOrFileAsync(string source, string destination)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperDirectoryActionResult resultObject = null;
-            try
-            {
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "source", source },
-                    { "dest", destination }
-                };
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "files/copy",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Post, $"files/copy", jsonObject: null, cts: default, urlSegments: urlSegments)
-                    .ConfigureAwait(false);
-                */
-                KlipperDirectoryActionRespone queryResult = GetObjectFromJson<KlipperDirectoryActionRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDirectoryActionRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<byte[]> DownloadFileAsync(string relativeFilePath)
-        {
-            try
-            {
-                //Uri uri = new($"{FullWebAddress}/server/files/{relativeFilePath}");
-                string uri = $"{FullWebAddress}/server/files/{relativeFilePath}";
-                byte[] file = await DownloadFileFromUriAsync(uri)
-                    .ConfigureAwait(false)
-                    ;
-                return file;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return null;
-            }
-        }
-
-        public async Task<KlipperFileActionResult> UploadFileAsync(string file, string root = "gcodes", string path = "", int timeout = 100000)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperFileActionResult resultObject = null;
-            try
-            {
-                result = await SendMultipartFormDataFileRestApiRequestAsync(file, root, path, timeout).ConfigureAwait(false);
-                KlipperFileActionResult queryResult = GetObjectFromJson<KlipperFileActionResult>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult;
-                //return result?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperFileActionResult),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return null;
-            }
-        }
-
-        public async Task<KlipperFileActionResult> UploadFileAsync(string fileName, byte[] file, string root = "gcodes", string path = "", int timeout = 10000)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperFileActionResult resultObject = null;
-            try
-            {
-                result = await SendMultipartFormDataFileRestApiRequestAsync(fileName, file, root, path, timeout).ConfigureAwait(false);
-                KlipperFileActionResult queryResult = GetObjectFromJson<KlipperFileActionResult>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult;
-                //return result?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperFileActionResult),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return null;
-            }
-        }
-
-        public async Task<KlipperDirectoryActionResult> DeleteFileAsync(string root, string filePath)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperDirectoryActionResult resultObject = null;
-            try
-            {
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Delete,
-                       command: $"files/{root}/{filePath}",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Delete, $"files/{root}/{filePath}")
-                    .ConfigureAwait(false);
-                */
-                KlipperDirectoryActionRespone queryResult = GetObjectFromJson<KlipperDirectoryActionRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDirectoryActionResult),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task<KlipperDirectoryActionResult> DeleteFileAsync(string filePath)
-        {
-            IRestApiRequestRespone result = null;
-            KlipperDirectoryActionResult resultObject = null;
-            if (string.IsNullOrEmpty(filePath))
-            {
-                return resultObject;
-            }
-            try
-            {
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Delete,
-                       command: $"files/{filePath}",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Delete, $"files/{filePath}")
-                    .ConfigureAwait(false);
-                */
-                KlipperDirectoryActionRespone queryResult = GetObjectFromJson<KlipperDirectoryActionRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDirectoryActionRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<byte[]> DownloadLogFileAsync(KlipperLogFileTypes logType)
-        {
-            try
-            {
-                string uri = $"{FullWebAddress}/server/files/{logType.ToString().ToLower()}.log";
-                byte[] file = await DownloadFileFromUriAsync(uri)
-                    .ConfigureAwait(false);
-                ;
-                return file;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return null;
-            }
-        }
-        #endregion
-
+      
         #region Authorization
 
         // Doc: https://moonraker.readthedocs.io/en/latest/web_api/#login-user
@@ -5027,706 +2587,7 @@ namespace AndreasReitberger.API.Moonraker
             }
         }
         #endregion
-
-        #region Database APIs
-        public async Task<List<string>> ListDatabaseNamespacesAsync()
-        {
-            IRestApiRequestRespone result = null;
-            List<string> resultObject = new();
-            try
-            {
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "database/list",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Get, $"database/list")
-                    .ConfigureAwait(false);
-                */
-                KlipperDatabaseNamespaceListRespone queryResult = GetObjectFromJson<KlipperDatabaseNamespaceListRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                return queryResult?.Result?.Namespaces ?? new();
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDatabaseNamespaceListRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshDatabaseNamespacesAsync()
-        {
-            try
-            {
-                List<string> result = await ListDatabaseNamespacesAsync().ConfigureAwait(false);
-                AvailableNamespaces = result ?? new();
-                if (AvailableNamespaces?.Count > 0)
-                {
-                    // Try to detect the current operating system
-                    OperatingSystem = AvailableNamespaces.Contains("mainsail") ?
-                        MoonrakerOperatingSystems.MainsailOS :
-                        MoonrakerOperatingSystems.FluiddPi
-                        ;
-                }
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                AvailableNamespaces = new();
-            }
-        }
-
-        public async Task<Dictionary<string, object>> GetDatabaseItemAsync(string namespaceName, string key = "", bool throwOnMissingNamespace = false)
-        {
-            IRestApiRequestRespone result = null;
-            Dictionary<string, object> resultObject = new();
-            try
-            {
-                if (AvailableNamespaces?.Count == 0 || AvailableNamespaces == null)
-                {
-                    AvailableNamespaces = await ListDatabaseNamespacesAsync().ConfigureAwait(false);
-                }
-                if (!AvailableNamespaces.Contains(namespaceName))
-                {
-                    if (throwOnMissingNamespace)
-                    {
-                        throw new ArgumentOutOfRangeException(namespaceName, "The requested namespace name was not found in the database!");
-                    }
-                    // If namespace is missing, just return an empty resultObject now
-                    else return resultObject;
-                }
-
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "namespace", namespaceName }
-                };
-                if (!string.IsNullOrEmpty(key)) urlSegments.Add("key", key);
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "database/item",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Get, $"database/item", default, null, urlSegments)
-                            .ConfigureAwait(false);
-                */
-                KlipperDatabaseItemRespone queryResult = GetObjectFromJson<KlipperDatabaseItemRespone>(result?.Result);
-                if (queryResult != null)
-                {
-                    resultObject = new()
-                    {
-                        { $"{namespaceName}{(!string.IsNullOrEmpty(key) ? $"|{key}" : "")}", queryResult?.Result?.Value }
-                    };
-                }
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDatabaseItemRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public new Task<ObservableCollection<IWebCamConfig>> GetWebCamConfigsAsync() => GetWebCamSettingsAsync();
-
-        public async Task<ObservableCollection<IWebCamConfig>> GetWebCamSettingsAsync()
-        {
-            string resultString = string.Empty;
-            IRestApiRequestRespone result = null;
-            ObservableCollection<IWebCamConfig> resultObject = new();
-            try
-            {
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Get,
-                       command: "webcams/list",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                KlipperWebcamConfigRespone configs = GetObjectFromJson<KlipperWebcamConfigRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                if (configs?.Result?.Webcams?.Count > 0)
-                    return new(configs?.Result?.Webcams ?? new());
-                else
-                    // If nothing is returned, try to get it from the database directly.
-                    return await GetWebCamSettingsFromDatabaseAsync().ConfigureAwait(false);
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDatabaseItemRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task<ObservableCollection<IWebCamConfig>> GetWebCamSettingsFromDatabaseAsync()
-        {
-            string resultString = string.Empty;
-            ObservableCollection<IWebCamConfig> resultObject = new();
-            try
-            {
-                // Both operating systems handles their datababase namespaces and keys differently....
-                // @fluidd
-                // It seems that the webcams setting are also stored in the namespace=webcams
-                //string currentNameSpace = OperatingSystem == MoonrakerOperatingSystems.MainsailOS ? "mainsail" : "webcams";
-                //string currentKey = OperatingSystem == MoonrakerOperatingSystems.MainsailOS ? "webcam" : "";
-
-                // Seems to be the write way for both, MainsailOS and Fluidd
-                string currentNameSpace = "webcams";
-                string currentKey = "";
-
-                Dictionary<string, object> result = await GetDatabaseItemAsync(currentNameSpace, currentKey).ConfigureAwait(false);
-                KeyValuePair<string, object>? pair = result?.FirstOrDefault();
-                if (string.IsNullOrEmpty(pair?.Key)) return resultObject;
-
-                //resultString = pair.Value.ToString();
-                resultString = pair.Value.Value.ToString();
-
-                switch (OperatingSystem)
-                {
-                    /*
-                    case MoonrakerOperatingSystems.MainsailOS:
-                        KlipperDatabaseMainsailValueWebcam mainsailObject = GetObjectFromJson<KlipperDatabaseMainsailValueWebcam>(resultString);
-                        if(mainsailObject?.Configs != null)
-                        {
-                            IEnumerable<KlipperDatabaseWebcamConfig> temp = mainsailObject.Configs.Select(item => new KlipperDatabaseWebcamConfig()
-                            {
-                                Enabled = true,
-                                Name = item.Name,
-                                Icon = item.Icon,
-                                FlipX = item.FlipX,
-                                FlipY = item.FlipY,
-                                Service = item.Service,
-                                TargetFps = item.TargetFps,
-                                Url = item.Url,
-                            });
-                            resultObject = new(temp);
-                        }
-                        break;
-                    */
-                    case MoonrakerOperatingSystems.MainsailOS:
-                    case MoonrakerOperatingSystems.FluiddPi:
-                        Dictionary<Guid, KlipperDatabaseFluiddValueWebcamConfig> fluiddObject = GetObjectFromJson<Dictionary<Guid, KlipperDatabaseFluiddValueWebcamConfig>>(resultString, NewtonsoftJsonSerializerSettings);
-                        if (fluiddObject?.Count > 0)
-                        {
-                            IEnumerable<KlipperDatabaseWebcamConfig> temp = fluiddObject.Select(item => new KlipperDatabaseWebcamConfig()
-                            {
-                                Id = item.Key,
-                                Enabled = item.Value.Enabled,
-                                Alias = item.Value.Name,
-                                FlipX = item.Value.FlipX,
-                                FlipY = item.Value.FlipY,
-                                Service = item.Value.Service,
-                                TargetFps = item.Value.Fpstarget,
-                                Url = item.Value.UrlStream,
-                                UrlSnapshot = item.Value.UrlSnapshot,
-                                Orientation = (long)item.Value.Rotation,
-                            });
-                            resultObject = new(temp);
-                        }
-                        break;
-                    default:
-                        break;
-                }
-
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = resultString,
-                    TargetType = nameof(KlipperDatabaseMainsailValueWebcam),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task RefreshWebCamConfigAsync()
-        {
-            try
-            {
-                ObservableCollection<IWebCamConfig> result = await GetWebCamSettingsAsync().ConfigureAwait(false);
-                WebCams = result ?? new();
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                WebCams = new();
-            }
-        }
-
-        public async Task<KlipperDatabaseSettingsGeneral> GetGeneralSettingsAsync()
-        {
-            string resultString = string.Empty;
-            KlipperDatabaseSettingsGeneral resultObject = null;
-            try
-            {
-                string currentNamespace = OperatingSystem == MoonrakerOperatingSystems.MainsailOS ? "mainsail" : "fluidd";
-                string currentKey = OperatingSystem == MoonrakerOperatingSystems.MainsailOS ? "general" : "uiSettings";
-
-                Dictionary<string, object> result = await GetDatabaseItemAsync(currentNamespace, currentKey).ConfigureAwait(false);
-
-                KeyValuePair<string, object>? pair = result?.FirstOrDefault();
-                if (string.IsNullOrEmpty(pair?.Key)) return resultObject;
-
-                //resultString = pair.Value.ToString();
-                resultString = pair.Value.Value.ToString();
-
-                switch (OperatingSystem)
-                {
-                    case MoonrakerOperatingSystems.MainsailOS:
-                        KlipperDatabaseMainsailValueGeneral mainsailObject = GetObjectFromJson<KlipperDatabaseMainsailValueGeneral>(resultString);
-                        if (mainsailObject != null)
-                        {
-                            resultObject = new()
-                            {
-                                DisplayCancelPrint = mainsailObject.DisplayCancelPrint,
-                                Printername = mainsailObject.Printername,
-                            };
-                        }
-                        break;
-                    case MoonrakerOperatingSystems.FluiddPi:
-                        KlipperDatabaseFluiddValueUiSettings fluiddObject = GetObjectFromJson<KlipperDatabaseFluiddValueUiSettings>(resultString);
-                        if (fluiddObject?.General != null)
-                        {
-                            resultObject = new()
-                            {
-                                Locale = fluiddObject.General.Locale,
-                                Printername = fluiddObject.General.InstanceName,
-                            };
-                        }
-                        break;
-                    default:
-                        break;
-                }
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = resultString,
-                    TargetType = nameof(KlipperDatabaseMainsailValueGeneral),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task RefreshGeneralSettingsAsync()
-        {
-            try
-            {
-                KlipperDatabaseSettingsGeneral result = await GetGeneralSettingsAsync().ConfigureAwait(false);
-                HostName = result?.Printername;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                HostName = string.Empty;
-            }
-        }
-
-        public async Task<List<KlipperDatabaseRemotePrinter>> GetRemotePrintersAsync()
-        {
-            string resultString = string.Empty;
-            List<KlipperDatabaseRemotePrinter> resultObject = new();
-            try
-            {
-                /*
-                if (OperatingSystem != MoonrakerOperatingSystems.MainsailOS)
-                {
-                    throw new NotSupportedException($"The method '{nameof(GetRemotePrintersAsync)}() is only support on '{MoonrakerOperatingSystems.MainsailOS}!");
-                }*/
-                string currentNamespace = OperatingSystem == MoonrakerOperatingSystems.MainsailOS ? "mainsail" : "fluidd";
-                string currentKey = OperatingSystem == MoonrakerOperatingSystems.MainsailOS ? "remote_printers" : "uiSettings";
-
-                Dictionary<string, object> result = await GetDatabaseItemAsync(currentNamespace, currentKey).ConfigureAwait(false);
-
-                KeyValuePair<string, object>? pair = result?.FirstOrDefault();
-                if (string.IsNullOrEmpty(pair?.Key)) return resultObject;
-
-                //resultString = pair.Value.ToString();
-                resultString = pair.Value.Value.ToString();
-
-                switch (OperatingSystem)
-                {
-                    case MoonrakerOperatingSystems.MainsailOS:
-                        List<KlipperDatabaseMainsailValueRemotePrinter> mainsailObject = GetObjectFromJson<List<KlipperDatabaseMainsailValueRemotePrinter>>(resultString);
-                        if (mainsailObject != null)
-                        {
-                            resultObject = new(mainsailObject.Select(item => new KlipperDatabaseRemotePrinter()
-                            {
-                                Hostname = item.Hostname,
-                                Port = item.Port,
-                                Settings = item.Settings,
-                                WebPort = item.WebPort,
-                            }));
-                        }
-                        break;
-                    case MoonrakerOperatingSystems.FluiddPi:
-#if DEBUG
-                        //throw new NotSupportedException($"The method '{nameof(GetRemotePrintersAsync)}() is only support on '{MoonrakerOperatingSystems.MainsailOS}!");
-#endif
-                    /*
-                    KlipperDatabaseFluiddValueUiSettings fluiddObject = GetObjectFromJson<KlipperDatabaseFluiddValueUiSettings>(resultString);
-                    if (fluiddObject?.General != null)
-                    {
-                        resultObject = new()
-                        {
-                            Locale = fluiddObject.General.Locale,
-                            Printername = fluiddObject.General.InstanceName,
-                        };
-                    }
-                    break;
-                    */
-                    default:
-                        break;
-                }
-
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = resultString,
-                    TargetType = nameof(KlipperDatabaseMainsailValueRemotePrinter),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task RefreshRemotePrintersAsync()
-        {
-            try
-            {
-                List<KlipperDatabaseRemotePrinter> result = await GetRemotePrintersAsync().ConfigureAwait(false);
-                Printers = new(result ?? new());
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Printers = new();
-            }
-        }
-
-        public async Task<List<KlipperDatabaseTemperaturePreset>> GetDashboardPresetsAsync()
-        {
-            string resultString = string.Empty;
-            List<KlipperDatabaseTemperaturePreset> resultObject = new();
-            try
-            {
-                string currentNamespace = OperatingSystem == MoonrakerOperatingSystems.MainsailOS ? "mainsail" : "fluidd";
-                string currentKey = OperatingSystem == MoonrakerOperatingSystems.MainsailOS ? "presets" : "uiSettings";
-
-                Dictionary<string, object> result = await GetDatabaseItemAsync(currentNamespace, currentKey).ConfigureAwait(false);
-                KeyValuePair<string, object>? pair = result?.FirstOrDefault();
-                if (string.IsNullOrEmpty(pair?.Key)) return resultObject;
-
-                resultString = pair.Value.Value.ToString();
-
-                switch (OperatingSystem)
-                {
-                    case MoonrakerOperatingSystems.MainsailOS:
-                        // New since latest update
-                        KlipperDatabaseMainsailValuePresets mainsailObject = GetObjectFromJson<KlipperDatabaseMainsailValuePresets>(resultString, NewtonsoftJsonSerializerSettings);
-                        //List<KlipperDatabaseMainsailValuePreset> mainsailObject = GetObjectFromJson<List<KlipperDatabaseMainsailValuePreset>>(resultString);
-                        if (mainsailObject != null)
-                        {
-                            IEnumerable<KlipperDatabaseTemperaturePreset> temp = mainsailObject.Presets.Select((item, index) => new KlipperDatabaseTemperaturePreset()
-                            {
-                                //Id = Guid.NewGuid(),
-                                Id = item.Key,
-                                Name = item.Value.Name,
-                                Gcode = item.Value.Gcode,
-                                Values = new(item.Value.Values.Select(valuePair => new KeyValuePair<string, KlipperDatabaseTemperaturePresetHeater>(
-                                    valuePair.Key, new KlipperDatabaseTemperaturePresetHeater()
-                                    {
-                                        Name = valuePair.Key,
-                                        Active = valuePair.Value.Bool,
-                                        Type = valuePair.Value.Type,
-                                        Value = valuePair.Value.Value,
-                                    }))),
-                            });
-                            resultObject = new(temp);
-                        }
-                        break;
-                    case MoonrakerOperatingSystems.FluiddPi:
-                        //resultString = pair.Value.ToString();
-                        KlipperDatabaseFluiddValueUiSettings fluiddObject = GetObjectFromJson<KlipperDatabaseFluiddValueUiSettings>(resultString);
-                        if (fluiddObject?.Dashboard?.TempPresets != null)
-                        {
-                            IEnumerable<KlipperDatabaseTemperaturePreset> temp = fluiddObject.Dashboard.TempPresets.Select(item => new KlipperDatabaseTemperaturePreset()
-                            {
-                                Id = item.Id,
-                                Name = item.Name,
-                                Gcode = item.Gcode,
-                                Values = new(item.Values.Select(valuePair => new KeyValuePair<string, KlipperDatabaseTemperaturePresetHeater>(
-                                    valuePair.Key, new KlipperDatabaseTemperaturePresetHeater()
-                                    {
-                                        Name = valuePair.Key,
-                                        Active = valuePair.Value.Active,
-                                        Type = valuePair.Value.Type,
-                                        Value = valuePair.Value.Value,
-                                    }))),
-                            });
-                            resultObject = new(temp);
-                        }
-                        break;
-                    default:
-                        break;
-                }
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = resultString,
-                    TargetType = nameof(KlipperDatabaseMainsailValuePreset),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshDashboardPresetsAsync()
-        {
-            try
-            {
-                // Could be null if no presets have been defined yet
-                List<KlipperDatabaseTemperaturePreset> result = await GetDashboardPresetsAsync().ConfigureAwait(false);
-                Presets = result ?? new();
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Presets = new();
-            }
-        }
-
-        public async Task<KlipperDatabaseMainsailValueHeightmapSettings> GetMeshHeightMapSettingsAsync()
-        {
-            string resultString = string.Empty;
-            KlipperDatabaseMainsailValueHeightmapSettings resultObject = null;
-            try
-            {
-                if (OperatingSystem != MoonrakerOperatingSystems.MainsailOS)
-                {
-                    throw new NotSupportedException($"The method '{nameof(GetMeshHeightMapSettingsAsync)}() is only support on '{MoonrakerOperatingSystems.MainsailOS}!");
-                }
-
-                Dictionary<string, object> result = await GetDatabaseItemAsync("mainsail", "heightmap").ConfigureAwait(false);
-                KeyValuePair<string, object>? pair = result?.FirstOrDefault();
-                if (string.IsNullOrEmpty(pair?.Key)) return resultObject;
-
-                //resultString = pair.Value.ToString();
-                resultString = pair.Value.Value.ToString();
-
-                resultObject = GetObjectFromJson<KlipperDatabaseMainsailValueHeightmapSettings>(resultString);
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = resultString,
-                    TargetType = nameof(KlipperDatabaseMainsailValueHeightmapSettings),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<Dictionary<string, object>> AddDatabaseItemAsync(string namespaceName, string key, object value)
-        {
-            IRestApiRequestRespone result = null;
-            Dictionary<string, object> resultObject = new();
-            try
-            {
-                object cmd = new
-                {
-                    @namespace = namespaceName,
-                    key = key,
-                    value = value,
-                };
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "database/item",
-                       jsonObject: cmd,
-                       authHeaders: AuthHeaders,
-                       //urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                //result = await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Post, "database/item", cmd).ConfigureAwait(false);
-                KlipperDatabaseItemRespone queryResult = GetObjectFromJson<KlipperDatabaseItemRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                if (queryResult != null)
-                {
-                    resultObject = new()
-                    {
-                        { $"{namespaceName}{(!string.IsNullOrEmpty(key) ? $"|{key}" : "")}", queryResult?.Result?.Value }
-                    };
-                }
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDatabaseItemRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-
-        public async Task<Dictionary<string, object>> DeleteDatabaseItemAsync(string namespaceName, string key)
-        {
-            IRestApiRequestRespone result = null;
-            Dictionary<string, object> resultObject = new();
-            try
-            {
-                Dictionary<string, string> urlSegments = new()
-                {
-                    { "namespace", namespaceName }
-                };
-                if (!string.IsNullOrEmpty(key)) urlSegments.Add("key", key);
-
-                string targetUri = $"{MoonrakerCommands.Server}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Delete,
-                       command: "database/item",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders,
-                       urlSegments: urlSegments,
-                       cts: default
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result =
-                    await SendRestApiRequestAsync(MoonrakerCommandBase.server, Method.Delete, $"database/item", default, null, urlSegments)
-                    .ConfigureAwait(false);
-                */
-                KlipperDatabaseItemRespone queryResult = GetObjectFromJson<KlipperDatabaseItemRespone>(result.Result, NewtonsoftJsonSerializerSettings);
-                if (queryResult != null)
-                {
-                    resultObject = new()
-                    {
-                        { $"{namespaceName}{(!string.IsNullOrEmpty(key) ? $"|{key}" : "")}", queryResult?.Result?.Value }
-                    };
-                }
-                return resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(KlipperDatabaseItemRespone),
-                    Message = jecx.Message,
-                });
-                return resultObject;
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        #endregion
-
+   
         #region Job Queue APIs
 
         public async Task<KlipperJobQueueResult> GetJobQueueStatusAsync()
